+++ src/main/java/bitnagil/bitnagil_backend/emotionMarble/service/EmotionMarbleService.java
@@ -5,29 +5,22 @@ import bitnagil.bitnagil_backend.emotionMarble.domain.enums.EmotionMarbleType;
 import bitnagil.bitnagil_backend.emotionMarble.repository.EmotionMarbleRepository;
 import bitnagil.bitnagil_backend.emotionMarble.request.RegisterEmotionMarbleRequest;
 import bitnagil.bitnagil_backend.emotionMarble.response.EmotionMarbleTypeResponse;
-import bitnagil.bitnagil_backend.emotionMarble.response.RecommendedRoutineDto;
-import bitnagil.bitnagil_backend.emotionMarble.response.RecommendedSubRoutineDto;
+import bitnagil.bitnagil_backend.recommendedRoutine.response.RecommendedRoutineDto;
 import bitnagil.bitnagil_backend.emotionMarble.response.RegisterEmotionMarbleResponse;
-import bitnagil.bitnagil_backend.global.entity.HistoryPk;
 import bitnagil.bitnagil_backend.global.errorcode.ErrorCode;
 import bitnagil.bitnagil_backend.global.exception.CustomException;
 
-import bitnagil.bitnagil_backend.onboarding.domain.Case;
-import bitnagil.bitnagil_backend.recommendedRoutine.domain.RecommendedRoutine;
-import bitnagil.bitnagil_backend.recommendedRoutine.domain.RecommendedSubRoutine;
-import bitnagil.bitnagil_backend.recommendedRoutine.repository.RecommendedRoutineRepository;
-import bitnagil.bitnagil_backend.recommendedRoutine.repository.RecommendedSubRoutineRepository;
+import bitnagil.bitnagil_backend.recommendedRoutine.service.RecommendedRoutineService;
 import bitnagil.bitnagil_backend.user.domain.User;
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 /**
  * 감정 구슬에 대한 로직을 관리하는 클래스입니다.
@@ -37,9 +30,8 @@ import java.util.UUID;
 public class EmotionMarbleService {
 
     private final EmotionMarbleRepository emotionMarbleRepository;
-    private final RecommendedRoutineRepository recommendRoutineRepository;
-    private final RecommendedSubRoutineRepository recommendedSubRoutineRepository;
 
+    private final RecommendedRoutineService recommendedRoutineService;
     private final EmotionMarbleFactory emotionMarbleFactory;
 
     // 감정 구술 조회(enum의 value를 가져온다.)
@@ -66,40 +58,13 @@ public class EmotionMarbleService {
         emotionMarbleRepository.save(emotionMarble);
 
         // 감정 구슬에 따른 추천 루틴 응답
-        List<RecommendedRoutine> recommendedRoutines = recommendRoutineRepository.findByResultCase(emotionMarble.getResultCase());
-        if (recommendedRoutines.isEmpty()) {
-            throw new CustomException(ErrorCode.NOT_FOUND_RECOMMENDED_ROUTINE);
-        }
+        List<RecommendedRoutineDto> recommendedRoutineDtoList =
+            recommendedRoutineService.recommendRoutinesByEmotionMarble(emotionMarble.getResultCase());
 
-        // 응답 생성
-        List<RecommendedRoutineDto> recommendedRoutineDtoList = new ArrayList<>();
-        for (RecommendedRoutine recommendedRoutine : recommendedRoutines) {
-            List<RecommendedSubRoutineDto> recommendedRoutineDetailDtoList = new ArrayList<>();
-
-            List<RecommendedSubRoutine> recommendedSubRoutines = recommendedSubRoutineRepository.findByRecommendedRoutine(recommendedRoutine);
-
-            // 추천 루틴의 세부 루틴을 dto로 변환한다.
-            for (RecommendedSubRoutine recommendedSubRoutine : recommendedSubRoutines) {
-                RecommendedSubRoutineDto recommendedRoutineDetailDto = RecommendedSubRoutineDto.builder()
-                        .recommendedSubRoutineId(recommendedSubRoutine.getRecommendedSubRoutineId())
-                        .recommendedSubRoutineName(recommendedSubRoutine.getSubRoutineName())
-                        .build();
-                recommendedRoutineDetailDtoList.add(recommendedRoutineDetailDto);
-            }
-
-            // 추천 루틴을 dto로 변환한다.
-            RecommendedRoutineDto recommendedRoutineDto = RecommendedRoutineDto.builder()
-                    .recommendedRoutineId(recommendedRoutine.getRecommendedRoutineId())
-                    .recommendedRoutineName(recommendedRoutine.getRecommendedRoutineName())
-                    .routineDescription(recommendedRoutine.getRecommendedRoutineDescription())
-                    .recommendedSubRoutines(recommendedRoutineDetailDtoList)
-                    .build();
-            recommendedRoutineDtoList.add(recommendedRoutineDto);
-        }
-
-        RegisterEmotionMarbleResponse response = RegisterEmotionMarbleResponse.builder()
+        return RegisterEmotionMarbleResponse.builder()
                 .recommendedRoutines(recommendedRoutineDtoList)
                 .build();
-        return response;
     }
+
+
 }
