+++ src/main/java/bitnagil/bitnagil_backend/onboarding/response/OnboardingResponse.java
@@ -1,5 +1,6 @@
 package bitnagil.bitnagil_backend.onboarding.response;
 
+import bitnagil.bitnagil_backend.recommendedRoutine.response.RecommendedRoutineDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
