+++ src/main/java/bitnagil/bitnagil_backend/recommendedRoutine/service/RecommendedRoutineMapper.java
@@ -1,0 +1,10 @@
+package bitnagil.bitnagil_backend.recommendedRoutine.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * 추천 루틴 관련해서 DB에서 조회해오거나 가공된 데이터를 DTO로 변환하는 Mapper 클래스입니다.
+ */
+@Service
+public class RecommendedRoutineMapper {
+}
