+++ src/main/java/bitnagil/bitnagil_backend/routine/service/RoutineMapper.java
@@ -14,7 +14,7 @@ import bitnagil.bitnagil_backend.routine.response.RoutineSearchResultDto;
 import bitnagil.bitnagil_backend.routine.response.SubRoutineSearchResultDto;
 
 /**
- * DTO로 변환하는 Mapper 클래스입니다.
+ * 루틴 관련해서 DB에서 조회해오거나 가공된 데이터를 DTO로 변환하는 Mapper 클래스입니다.
  */
 @Service
 public class RoutineMapper {
